fix : 채팅 입력 시 마지막 채팅이 가려지는 문제 수정
