+++ src/components/ChatroomDetailContents/ChatroomDetailContents.style.tsx
@@ -4,7 +4,7 @@ export const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
-  padding-bottom: 0;
+  flex-grow: 1;
 `;
 
 export const Date = styled.div`
@@ -14,4 +14,8 @@ export const Date = styled.div`
     text-align: center;
     color: ${colors.grey4};
   `}
+`;
+
+export const LastBottomBlock = styled.span`
+  height: 3.5rem;
 `;
